refactored validation into seperate module
