+++ src/main.py
@@ -1,19 +1,5 @@
 import sys
-import os
-import zipfile
-
-def check_zip_file(file):
-
-    # first ensure file exists
-    if not os.path.exists(file):
-        return f"Error: file '{file}' does not exist"
-
-    # then check if its a zip
-    if zipfile.is_zipfile(file):
-        return f"{file} is a zip file"
-    else:
-        return f"{file} is not a zip file"
-
+from src.validator import zipvalidation
 
 def main():
     # get the file as a command-line argument
@@ -28,7 +14,7 @@ def main():
         return
 
     file = sys.argv[1]
-    result = check_zip_file(file)
+    result = zipvalidation.check_zip_file(file)
     print(result)
     print("\nMain working correctly.\n")
